backend/requirements.txt
# ============================================
# Sistema de Gestión del Edificio Multifuncional
# ============================================

# ------------------------
# Framework web
# ------------------------
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==4.0.0

# ------------------------
# Autenticación y seguridad
# ------------------------
Flask-JWT-Extended==4.7.1
bcrypt==4.2.1
Authlib==1.3.1
requests-oauthlib==2.0.0

# ------------------------
# Base de datos
# ------------------------
# Para PostgreSQL en Windows/Linux
psycopg2-binary==2.9.10
# Si usas SQL Server, instala también: pyodbc

# ------------------------
# Validación y serialización
# ------------------------
marshmallow==3.23.1
flask-marshmallow==1.2.1
marshmallow-sqlalchemy==1.1.0

# ------------------------
# Documentación API
# ------------------------
flasgger==0.9.7.1

# ------------------------
# Configuración y utilidades
# ------------------------
python-dotenv==1.0.1
requests==2.32.3
